Add encoding parameter to parsing functions
Prior function assumed utf8 data files, but the code was only written to
work on files with one-byte characters.

Currently, the code only works with single-byte character encodings
because it uses struct.Struct to parse records. It could be changed to
work with multi-byte encodings (like utf8) by, for example, using slices
to split up fields. That would be slower, however.

A good question is whether there are many fixed width files in the wild
that use multi-byte characters.