Speed up parsing of yes/no values
Use a frozenset to store valid 'yes' and 'no' values in order to speed
up lookups. Move type check to end. It is faster to assume values passed
in are valid and see what went wrong at the end (possibly a type error)
than to confirm value is string or bytes upfront.